size_t⇒DWORDのキャストを追加
64bitビルドでの警告を抑制するため。

+++ sakura_core/view/CTextDrawer.cpp
@@ -312,7 +312,7 @@ void CTextDrawer::DispNoteLines(
 		}
 
 		std::vector<DWORD> vNumPts(vLineEnds.size() / 2, 2);
-		::PolyPolyline(gr, vLineEnds.data(), vNumPts.data(), vNumPts.size());
+		::PolyPolyline(gr, vLineEnds.data(), vNumPts.data(), static_cast<DWORD>(vNumPts.size()));
 	}
 }
